refactor: change require to import add types in board module

+++ src/resources/boards/board.service.ts
@@ -1,25 +1,26 @@
-const boardsRepo = require('./board.memory.repository');
-const tasksService = require('../tasks/task.service');
+import boardsRepo from './board.memory.repository';
+import tasksService from '../tasks/task.service';
+import { IBoard, IBoardProps } from './board.model';
 
 /**
  * Get all boards from db
  * @returns {Promise<Array<Board>>} returns promise with array of all boards
  */
-const getAll = () => boardsRepo.getAll();
+const getAll = (): Promise<IBoard[]> => boardsRepo.getAll();
 
 /**
  * Get single board from db
  * @param {string} id id of single board
  * @returns {Promise<Board>} returns promise with single board
  */
-const get = (id) => boardsRepo.get(id);
+const get = (id: string): Promise<IBoard> => boardsRepo.get(id);
 
 /**
  * Request to db to create a new board
  * @param {Board} body board to be created
  * @returns {Promise<Board>} returns promise of created board
  */
-const create = (body) => boardsRepo.create(body);
+const create = (body: IBoardProps): Promise<IBoard> => boardsRepo.create(body);
 
 /**
  * Update board
@@ -27,7 +28,8 @@ const create = (body) => boardsRepo.create(body);
  * @param {Board} body object with data to update
  * @returns {Promise<Board>} returns promise with updated board
  */
-const update = (id, body) => boardsRepo.update(id, body);
+const update = (id: string, body: IBoard): Promise<IBoard> =>
+  boardsRepo.update(id, body);
 
 /**
  * Delete board from db
@@ -35,9 +37,11 @@ const update = (id, body) => boardsRepo.update(id, body);
  * @param {string} id id of board to be deleted
  * @returns {Promise<{}>} returns promise with an empty object if board deleted
  */
-const deleteBoard = async(id) => {
-  await boardsRepo.deleteBoard(id);
-  await tasksService.deleteAllTasksBelongsToBoard(id);
+const deleteBoard = async (id: string): Promise<void> => {
+  const isBoardDeleted = await boardsRepo.deleteBoard(id);
+  if (isBoardDeleted) {
+    await tasksService.deleteAllTasksBelongsToBoard(id);
+  }
 };
 
-module.exports = {getAll, get, create, update, deleteBoard};
+export default { getAll, get, create, update, deleteBoard };
